+++ src/array.h
@@ -11,11 +11,11 @@ struct Array {
 
     void reserve(u32 amount);
     void add(T item);
-    void remove(u32 index);
-    T* get(u32 index);
+    void remove(u64 index);
+    T* get(u64 index);
 
     // This will reorder the array!
-    void fast_remove(u32 index);
+    void fast_remove(u64 index);
 };
 
 #endif
